requirements.txt
opencv_python>=4.3
mlflow==1.15.0
matplotlib==3.4.1
pandas==1.1.5
requests==2.25.1
plotly==5.3.1
torchvision==0.10
torch==1.9
optuna==2.9.1
pytorch-lightning>=1.4.7
numpy==1.20.2
scipy==1.6.2
tqdm==4.60.0
Pillow==8.3.2
PyYAML>=5.3
scikit_learn==0.24.2
seaborn>=0.10
